chore: move test to same dir as unit under test
